requirements.txt
altgraph==0.17
progressbar2==3.53.1
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
python-utils==2.5.6
pytube==10.7.2
six==1.15.0
